backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
redis>=5.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
sentry-sdk[fastapi]>=2.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0
